Regel-refaktoring: Gruppere regler i packages og fjerne ReglerForGrunnforordningen.
