+++ src/main/kotlin/no/nav/medlemskap/regler/v1/RegelFactory.kt
@@ -7,6 +7,9 @@ import no.nav.medlemskap.regler.common.RegelId.*
 import no.nav.medlemskap.regler.v1.arbeidsforhold.HarBrukerSammenhengendeArbeidsforholdRegel
 import no.nav.medlemskap.regler.v1.arbeidsforhold.HarForetaketMerEnn5AnsatteRegel
 import no.nav.medlemskap.regler.v1.lovvalg.*
+import no.nav.medlemskap.regler.v1.medlemskap.*
+import no.nav.medlemskap.regler.v1.registrerteOpplysninger.ErBrukerEøsBorgerRegel
+import no.nav.medlemskap.regler.v1.registrerteOpplysninger.HarBrukerRegistrerteOpplysningerRegel
 
 class RegelFactory(private val datagrunnlag: Datagrunnlag) {
