+++ src/main/kotlin/no/nav/medlemskap/regler/v1/medlemskap/ErBrukersDekningUavklartRegel.kt
@@ -1,4 +1,4 @@
-package no.nav.medlemskap.regler.v1
+package no.nav.medlemskap.regler.v1.medlemskap
 
 import no.nav.medlemskap.domene.*
 import no.nav.medlemskap.domene.Dekning.Companion.uavklartForYtelse
@@ -12,7 +12,7 @@ class ErBrukersDekningUavklartRegel(
         ytelse: Ytelse,
         val medlemskap: List<Medlemskap>,
         val periode: InputPeriode
-) : MedlemRegel(RegelId.REGEL_1_6, ytelse, periode, medlemskap) {
+) : MedlemskapRegel(RegelId.REGEL_1_6, ytelse, periode, medlemskap) {
 
     override fun operasjon(): Resultat {
         val gjeldendeDekning = medlemskap.gjeldendeDekning(kontrollPeriodeForMedl)
