+++ src/main/kotlin/no/nav/medlemskap/regler/v1/medlemskap/ErPeriodeMedMedlemskapInnenfor12MndPeriodeRegel.kt
@@ -1,4 +1,4 @@
-package no.nav.medlemskap.regler.v1
+package no.nav.medlemskap.regler.v1.medlemskap
 
 import no.nav.medlemskap.domene.Datagrunnlag
 import no.nav.medlemskap.domene.InputPeriode
@@ -7,19 +7,19 @@ import no.nav.medlemskap.domene.Ytelse
 import no.nav.medlemskap.regler.common.RegelId
 import no.nav.medlemskap.regler.common.Resultat
 
-class ErPeriodeUtenMedlemskapInnenfor12MndPeriodeRegel(
+class ErPeriodeMedMedlemskapInnenfor12MndPeriodeRegel(
         ytelse: Ytelse,
         periode: InputPeriode,
         medlemskap: List<Medlemskap>
-) : MedlemRegel(RegelId.REGEL_1_3_1, ytelse, periode, medlemskap) {
+) : MedlemskapRegel(RegelId.REGEL_1_4, ytelse, periode, medlemskap) {
 
     override fun operasjon(): Resultat {
-        return erMedlemskapPeriodeOver12MndPeriode(false)
+        return erMedlemskapPeriodeOver12MndPeriode(true)
     }
 
     companion object {
-        fun fraDatagrunnlag(datagrunnlag: Datagrunnlag): ErPeriodeUtenMedlemskapInnenfor12MndPeriodeRegel {
-            return ErPeriodeUtenMedlemskapInnenfor12MndPeriodeRegel(
+        fun fraDatagrunnlag(datagrunnlag: Datagrunnlag): ErPeriodeMedMedlemskapInnenfor12MndPeriodeRegel {
+            return ErPeriodeMedMedlemskapInnenfor12MndPeriodeRegel(
                     ytelse = datagrunnlag.ytelse,
                     periode = datagrunnlag.periode,
                     medlemskap = datagrunnlag.medlemskap
