+++ src/main/kotlin/no/nav/medlemskap/regler/v1/medlemskap/PeriodeMedOgUtenMedlemskapRegel.kt
@@ -1,4 +1,4 @@
-package no.nav.medlemskap.regler.v1
+package no.nav.medlemskap.regler.v1.medlemskap
 
 import no.nav.medlemskap.domene.Datagrunnlag
 import no.nav.medlemskap.domene.InputPeriode
@@ -8,25 +8,25 @@ import no.nav.medlemskap.regler.common.RegelId
 import no.nav.medlemskap.regler.common.Resultat
 import no.nav.medlemskap.regler.common.ja
 import no.nav.medlemskap.regler.common.nei
-import no.nav.medlemskap.regler.funksjoner.MedlFunksjoner.harPeriodeMedMedlemskap
+import no.nav.medlemskap.regler.funksjoner.MedlFunksjoner.harMedlPeriodeMedOgUtenMedlemskap
 
-class PeriodeMedMedlemskapRegel(
+class PeriodeMedOgUtenMedlemskapRegel(
         ytelse: Ytelse,
         private val periode: InputPeriode,
         private val medlemskap: List<Medlemskap>
-) : MedlemRegel(RegelId.REGEL_1_3, ytelse, periode, medlemskap) {
+) : MedlemskapRegel(RegelId.REGEL_1_2, ytelse, periode, medlemskap) {
 
     override fun operasjon(): Resultat {
         return when {
-            medlemskap harPeriodeMedMedlemskap kontrollPeriodeForMedl -> ja()
+            medlemskap harMedlPeriodeMedOgUtenMedlemskap kontrollPeriodeForMedl -> ja()
             else -> nei()
         }
     }
 
     companion object {
 
-        fun fraDatagrunnlag(datagrunnlag: Datagrunnlag): PeriodeMedMedlemskapRegel {
-            return PeriodeMedMedlemskapRegel(
+        fun fraDatagrunnlag(datagrunnlag: Datagrunnlag): PeriodeMedOgUtenMedlemskapRegel {
+            return PeriodeMedOgUtenMedlemskapRegel(
                     ytelse = datagrunnlag.ytelse,
                     periode = datagrunnlag.periode,
                     medlemskap = datagrunnlag.medlemskap
