+++ src/main/kotlin/no/nav/medlemskap/regler/v1/registrerteOpplysninger/FinnesOpplysningerIMedlRegel.kt
@@ -1,4 +1,4 @@
-package no.nav.medlemskap.regler.v1
+package no.nav.medlemskap.regler.v1.registrerteOpplysninger
 
 import no.nav.medlemskap.common.medlCounter
 import no.nav.medlemskap.domene.Datagrunnlag
